Update non-audit / non-pentest note
The situation is still the same, the code has neither been audited, nor
been pentested and it's in the responsibility of the user to check the
code before deciding whether it is safe / trustworthy to use it or not.

To not scare/confuse users the formulation was changed. As other open-source
software, the code was developed to the best of knowledge and belief.